backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
Pillow>=10.0.0
ffmpeg-python==0.2.0
pydub==0.25.1
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles>=24.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
